% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ke_networks.R
\name{compute_paths_for_experiment}
\alias{compute_paths_for_experiment}
\title{compute_paths_for_experiment}
\usage{
compute_paths_for_experiment(
  expi,
  g_main,
  ke_list_exp,
  docking_list,
  Enrichment_data_BMD,
  ke_all_deg_list,
  target_node = "Increased, Liver Steatosis",
  normalize = NULL,
  PBK_time = 40
)
}
\value{
Returns list of paths to target node in network.
}
\description{
Path computation (unchanged logic, including the original pp/mx behavior)
}
