% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ke_networks.R
\name{get_ke_net_reverse_dosimetry}
\alias{get_ke_net_reverse_dosimetry}
\title{Build network/KE lists per experiment}
\usage{
get_ke_net_reverse_dosimetry(
  reverse_dosimetry_summarized,
  max_path_length = 2,
  n_AOs = 1,
  n_MIEs = 1,
  mode = "out",
  enlarge_ke_selection = TRUE,
  ke_id = "ke",
  group_by = "aop",
  normalize = NULL
)
}
\arguments{
\item{max_path_length}{Integer; default 2.}

\item{n_AOs}{Integer; default 1.}

\item{n_MIEs}{Integer; default 1.}

\item{enlarge_ke_selection}{Logical; default TRUE.}

\item{ke_id}{Character; default "ke".}

\item{group_by}{Character; group variable for plotting; default "aop".}

\item{reverse}{dosimetry summarized results per experiment.}

\item{Biological_system_annotations}{Data frame with column: key_event_name (and other metadata used by make_visNetwork).}

\item{experiments}{Optional character vector of experiment IDs; default: names(probability_matrices).}
}
\value{
A named list with elements:
\itemize{
  \item net: per-experiment nodes/edges from make_visNetwork
  \item visnet: per-experiment visNetwork objects
}
}
\description{
Build network/KE lists per experiment
}
