% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ke_networks.R
\name{score_path}
\alias{score_path}
\title{score_path}
\usage{
score_path(
  path_nodes,
  DD,
  DEG,
  BMD,
  PBK,
  DEG_score,
  Binding,
  w_BMD = 2,
  w_PBK = 1.5,
  w_DEG = 1,
  w_Binding = 0.5,
  alpha = 0.2,
  normalize = NULL
)
}
\value{
Returns list of normalized scores.
}
\description{
Simple path support scoring function
}
